+++ heat/core/tests/test_constants.py
@@ -1,9 +1,6 @@
-import unittest
 import numpy as np
-import os
 
 import heat as ht
-
 from .test_suites.basic_test import TestCase