+++ heat/core/tests/test_operations.py
@@ -1,8 +1,6 @@
-import unittest
 import torch
-import os
-import heat as ht
 
+import heat as ht
 from .test_suites.basic_test import TestCase