+++ heat/core/tests/test_rounding.py
@@ -1,8 +1,7 @@
-import torch
-import unittest
 import numpy as np
+import torch
+
 import heat as ht
-import os
 from .test_suites.basic_test import TestCase