+++ heat/core/tests/test_statistics.py
@@ -1,11 +1,10 @@
 import numpy as np
 import torch
-import unittest
-from itertools import combinations
-import os
-from scipy import stats as ss
-import heat as ht
 
+from itertools import combinations
+from scipy import stats as ss
+
+import heat as ht
 from .test_suites.basic_test import TestCase