+++ heat/core/tests/test_suites/test_basic_test.py
@@ -1,8 +1,7 @@
-import os
-import heat as ht
 import numpy as np
 import torch
 
+import heat as ht
 from .basic_test import TestCase