+++ heat/core/tests/test_types.py
@@ -1,9 +1,7 @@
 import numpy as np
 import torch
-import unittest
-import os
-import heat as ht
 
+import heat as ht
 from .test_suites.basic_test import TestCase